User database operations moved into own file.
